requirements.chat.txt
fastapi
uvicorn
websockets
redis
psycopg2-binary